requirements.txt
locust==2.27.0
pytest==8.1.1
fastapi==0.111.0
fastapi-cli==0.0.3
pyodbc==5.1.0
python-dotenv==1.0.1